Remove `XLAOptFuncWrapper` and replace with `XLAOptCallback`
The package now uses an `XLAOptCallback` to replace the optimizer with
an `XLAOptimProxy` instance which calls `xm.optimizer_step` during
the `opt.step`

The XLAOptCallback adds the proxy during the `begin_fit` event
and removes it during `after_fit` event.

This is more in line with the fastai idioms to override training
behavior using callbacks.